Add custom exception
Adds a custom exception to be used when we want to make an
ApplicationRecord (database) model's attribute read-only. This is useful
when we have a model like Roaster, where multiple users could be
consuming the model/data. We don't want something like a Roaster's name
to change once set, as this could have unexpected consequences.